+++ Source/Alternet.UI/Controls/Labels/GenericLabel.cs
@@ -14,12 +14,19 @@ namespace Alternet.UI
     public partial class GenericLabel : UserPaintControl
     {
         private string text = string.Empty;
+        private Color? textBackColor;
+        private bool imageVisible = true;
+        private int mnemonicCharIndex = -1;
+        private GenericAlignment alignment = GenericAlignment.TopLeft;
+        private string? textPrefix;
+        private string? textSuffix;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="GenericLabel"/> class.
         /// </summary>
         /// <param name="text">Value of the <see cref="Text"/> property.</param>
         public GenericLabel(string text)
+            : this()
         {
             this.text = text ?? string.Empty;
         }
@@ -29,6 +36,184 @@ namespace Alternet.UI
         /// </summary>
         public GenericLabel()
         {
+            BehaviorOptions = ControlOptions.DrawDefaultBackground
+                | ControlOptions.DrawDefaultBorder
+                | ControlOptions.RefreshOnCurrentState;
+        }
+
+        /// <summary>
+        /// Occurs when the <see cref="Image"/> property changes.
+        /// </summary>
+        public event EventHandler? ImageChanged;
+
+        /// <summary>
+        /// Gets or sets text prefix.
+        /// </summary>
+        /// <remarks>
+        /// Value of this property is shown at the beginning of <see cref="Text"/>
+        /// when control is painted.
+        /// </remarks>
+        public virtual string? TextPrefix
+        {
+            get
+            {
+                return textPrefix;
+            }
+
+            set
+            {
+                if (textPrefix == value)
+                    return;
+                textPrefix = value;
+                Invalidate();
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets text suffix.
+        /// </summary>
+        /// <remarks>
+        /// Value of this property is shown at the end of <see cref="Text"/>
+        /// when control is painted.
+        /// </remarks>
+        public string? TextSuffix
+        {
+            get
+            {
+                return textSuffix;
+            }
+
+            set
+            {
+                if (textSuffix == value)
+                    return;
+                textSuffix = value;
+                Invalidate();
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets background color of the text.
+        /// </summary>
+        /// <remarks>
+        /// By default is <c>null</c>. It means now additional background is painted
+        /// under the text.
+        /// </remarks>
+        [DefaultValue(null)]
+        public virtual Color? TextBackColor
+        {
+            get
+            {
+                return textBackColor;
+            }
+
+            set
+            {
+                if (textBackColor == value)
+                    return;
+                textBackColor = value;
+                Invalidate();
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets alignment of the text.
+        /// </summary>
+        public GenericAlignment TextAlignment
+        {
+            get
+            {
+                return alignment;
+            }
+
+            set
+            {
+                if (alignment == value)
+                    return;
+                alignment = value;
+                Invalidate();
+            }
+        }
+
+        /// <inheritdoc/>
+        public override Thickness Padding
+        {
+            get => base.Padding;
+            set
+            {
+                if (Padding == value)
+                    return;
+                base.Padding = value;
+                Invalidate();
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets underlined character index.
+        /// </summary>
+        /// <remarks>
+        /// By default equals -1.
+        /// </remarks>
+        [DefaultValue(null)]
+        public virtual int MnemonicCharIndex
+        {
+            get
+            {
+                return mnemonicCharIndex;
+            }
+
+            set
+            {
+                if (mnemonicCharIndex == value)
+                    return;
+                mnemonicCharIndex = value;
+                Invalidate();
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets the image that is displayed near the text.
+        /// </summary>
+        [DefaultValue(null)]
+        public virtual Image? Image
+        {
+            get
+            {
+                return StateObjects?.Images?.GetObjectOrNull(GenericControlState.Normal);
+            }
+
+            set
+            {
+                if (Image == value)
+                    return;
+                StateObjects ??= new();
+                StateObjects.Images ??= new();
+                StateObjects.Images.Normal = value;
+                RaiseImageChanged(EventArgs.Empty);
+                Invalidate();
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets the disabled image that is displayed near the text.
+        /// </summary>
+        public virtual Image? DisabledImage
+        {
+            get
+            {
+                return StateObjects?.Images?.GetObjectOrNull(GenericControlState.Disabled);
+            }
+
+            set
+            {
+                if (DisabledImage == value)
+                    return;
+                StateObjects ??= new();
+                StateObjects.Images ??= new();
+                StateObjects.Images.Disabled = value;
+                RaiseImageChanged(EventArgs.Empty);
+                Invalidate();
+            }
         }
 
         /// <inheritdoc/>
@@ -51,19 +236,50 @@ namespace Alternet.UI
             }
         }
 
+        /// <summary>
+        /// Gets or sets a value indicating whether to draw image.
+        /// </summary>
+        public virtual bool ImageVisible
+        {
+            get
+            {
+                return imageVisible;
+            }
+
+            set
+            {
+                if (imageVisible == value)
+                    return;
+                imageVisible = value;
+                Invalidate();
+            }
+        }
+
+        /// <summary>
+        /// Raises the <see cref="ImageChanged"/> event and calls
+        /// <see cref="OnImageChanged(EventArgs)"/>.
+        /// </summary>
+        /// <param name="e">An <see cref="EventArgs"/> that contains the event
+        /// data.</param>
+        public void RaiseImageChanged(EventArgs e)
+        {
+            OnImageChanged(e);
+            ImageChanged?.Invoke(this, e);
+        }
+
         /// <inheritdoc/>
         public override SizeD GetPreferredSize(SizeD availableSize)
         {
             var specifiedWidth = SuggestedWidth;
             var specifiedHeight = SuggestedHeight;
 
-            SizeD result = SizeD.Empty;
+            SizeD result = 0;
 
             var text = Text;
             if (text is not null)
             {
                 using var dc = CreateDrawingContext();
-                result = dc.GetTextExtent(text, GetLabelFont(), this);
+                result = dc.GetTextExtent(text, GetLabelFont(GenericControlState.Normal), this);
             }
 
             if (!double.IsNaN(specifiedWidth))
@@ -82,15 +298,104 @@ namespace Alternet.UI
         }
 
         /// <summary>
+        /// Called when the value of the <see cref="Image"/> property changes.
+        /// </summary>
+        /// <param name="e">An <see cref="EventArgs"/> that contains the event
+        /// data.</param>
+        protected virtual void OnImageChanged(EventArgs e)
+        {
+        }
+
+        /// <summary>
         /// Gets <see cref="Font"/> which is used to draw labels text.
         /// </summary>
         /// <returns></returns>
-        protected virtual Font GetLabelFont()
+        protected virtual Font GetLabelFont(GenericControlState state)
         {
-            var result = Font ?? UI.Control.DefaultFont;
+            var font = StateObjects?.Colors?.GetObjectOrNull(state)?.Font;
+
+            var result = font ?? Font ?? UI.Control.DefaultFont;
             if (IsBold)
                 result = result.AsBold;
             return result;
+        }
+
+        /// <summary>
+        /// Gets <see cref="Color"/> which is used to draw background of the label text.
+        /// </summary>
+        /// <returns>By default returns <see cref="Color.Empty"/> which means do not
+        /// draw background under the label text. In this case control's background
+        /// is used.</returns>
+        protected virtual Color GetLabelBackColor(GenericControlState state)
+        {
+            var color = StateObjects?.Colors?.GetObjectOrNull(state)?.BackgroundColor;
+
+            return color ?? TextBackColor ?? Color.Empty;
+        }
+
+        /// <inheritdoc/>
+        protected override void OnCurrentStateChanged(EventArgs e)
+        {
+            base.OnCurrentStateChanged(e);
+
+            if (StateObjects is null)
+                return;
+
+            if (StateObjects.HasOtherBackgrounds || StateObjects.HasOtherImages
+                || StateObjects.HasOtherBorders || StateObjects.HasOtherColors)
+                Refresh();
+        }
+
+        /// <summary>
+        /// Gets <see cref="Color"/> which is used to draw label text.
+        /// </summary>
+        /// <returns></returns>
+        protected virtual Color GetLabelForeColor(GenericControlState state)
+        {
+            var color = StateObjects?.Colors?.GetObjectOrNull(state)?.ForegroundColor;
+
+            if(color is null)
+            {
+                if (Enabled)
+                    color = ForeColor;
+                else
+                    color = SystemColors.GrayText;
+            }
+
+            return color;
+        }
+
+        /// <inheritdoc/>
+        protected override void DefaultPaint(Graphics dc, RectD rect)
+        {
+            BeforePaint(dc, rect);
+
+            var state = CurrentState;
+            var paddedRect = (
+                rect.Location + Padding.LeftTop,
+                rect.Size - Padding.Size);
+
+            DrawDefaultBackground(dc, rect);
+
+            if (Text != null)
+            {
+                var image = StateObjects?.Images?.GetObjectOrNull(state);
+                image ??= Image;
+                var imageOverride = ImageVisible ? image : null;
+                var textOverride = $"{TextPrefix}{Text}{TextSuffix}";
+
+                dc.DrawLabel(
+                    textOverride,
+                    GetLabelFont(state),
+                    GetLabelForeColor(state),
+                    GetLabelBackColor(state),
+                    imageOverride,
+                    paddedRect,
+                    TextAlignment,
+                    MnemonicCharIndex);
+            }
+
+            AfterPaint(dc, rect);
         }
 
         internal class LabelHandler : ControlHandler<GenericLabel>
@@ -99,22 +404,16 @@ namespace Alternet.UI
 
             public override void OnPaint(Graphics drawingContext)
             {
-                if (Control.Text != null)
+                Control.DefaultPaint(drawingContext, Control.DrawClientRectangle);
+            }
+
+            internal override Native.Control CreateNativeControl()
+            {
+                var result = new Native.Panel
                 {
-                    Color color;
-
-                    if (Control.Enabled)
-                        color = Control.ForeColor;
-                    else
-                        color = SystemColors.GrayText;
-
-                    drawingContext.DrawText(
-                        Control.Text,
-                        Control.ChildrenLayoutBounds.Location,
-                        Control.GetLabelFont(),
-                        color,
-                        Color.Empty);
-                }
+                    AcceptsFocusAll = false,
+                };
+                return result;
             }
         }
     }